requirements.txt
Django==2.2.10
mysqlclient
django-mysql
django-rest-framework